+++ AWS-Public/aws-public-ip-address-ranges.py
@@ -1,9 +1,11 @@
 #!/usr/bin/env python
+import ipaddress
 import json
 
 import click
 import requests
-from helper.selector import prompt_multi_selection
+from InquirerPy import inquirer
+from InquirerPy.base.control import Choice
 
 IP_TYPES = {
     "ipv4": ("ip_prefix", "prefixes"),
@@ -11,6 +13,23 @@ IP_TYPES = {
 }
 
 SOURCE = "https://ip-ranges.amazonaws.com/ip-ranges.json"
+
+
+def prompt_multi_selection(name, options, pre_selected_options, message=None):
+    if not options:
+        raise ValueError("No options retrieved for selection.")
+
+    choices = [
+        Choice(option, enabled=option in pre_selected_options)
+        for option in options
+    ]
+
+    return inquirer.checkbox(
+        message=message if message else f"Please choose the {name}",
+        choices=choices,
+        cycle=True,
+        transformer=lambda result: f"{len(result)} {name} selected",
+    ).execute()
 
 
 def list_ip_ranges(data, services, ip_type, region, prefix):
@@ -36,14 +55,29 @@ def select_services(data, ip_type):
     return prompt_multi_selection("Service", options=services, pre_selected_options=[])
 
 
-@click.command(help="Get AWS public IP ranges")
+def find_ip(ip, data, ip_type):
+    key_prefix, key_prefixes = IP_TYPES[ip_type]
+    target_ip = ipaddress.ip_address(ip)
+    for item in data[key_prefixes]:
+        if target_ip in ipaddress.ip_network(item[key_prefix]):
+            print(f"Found {ip} in")
+            print(json.dumps(item, indent=2))
+            return True
+
+
+@click.command(help="Get AWS public IP ranges, or check if a given IP is within AWS public IP ranges")
+@click.option("--ip", "-i", help="IP address to search.")
 @click.option("--prefix", "-p", help="Find the specified prefix e.g. 52.65.0.0/16 or 52.65.0.0; all if not specified")
 @click.option("--region", "-r", help="AWS Region; not specified for all regions", default="ap-southeast-2")
 @click.option("--ip-type", "-t", type=click.Choice(IP_TYPES, case_sensitive=False), default="ipv4", show_default=True)
 @click.option("--services", "-s", multiple=True)
-def main(prefix, region, ip_type, services):
+def main(ip, prefix, region, ip_type, services):
     data = requests.get(SOURCE).json()
     ip_type = ip_type.lower()
+
+    if ip:
+        find_ip(ip, data, ip_type)
+        return
 
     if services:
         services = list(map(str.upper, services))